+++ packages/conventional-recommended-bump/index.js
@@ -29,7 +29,12 @@ function conventionalRecommendedBump (optionsArgument, parserOptsArgument, cbArg
     try {
       presetPackage = conventionalChangelogPresetLoader(options.preset)
     } catch (err) {
-      return cb(new Error(`Unable to load the "${options.preset}" preset package. Please make sure it's installed.`))
+      if (err.message === 'does not exist') {
+        const preset = typeof options.preset === 'object' ? options.preset.name : options.preset
+        return cb(new Error(`Unable to load the "${preset}" preset package. Please make sure it's installed.`))
+      } else {
+        return cb(err)
+      }
     }
   }
 
@@ -77,7 +82,7 @@ function conventionalRecommendedBump (optionsArgument, parserOptsArgument, cbArg
             warn(`No commits since last release`)
           }
 
-          let result = whatBump(commits)
+          let result = whatBump(commits, options)
 
           if (result && result.level != null) {
             result.releaseType = VERSIONS[result.level]